+++ TMessagesProj/jni/voip/webrtc/modules/audio_processing/agc2/rnn_vad/rnn.h
@@ -29,19 +29,19 @@ namespace rnn_vad {
 // over-allocate space for fully-connected layers output vectors (implemented as
 // std::array). The value should equal the number of units of the largest
 // fully-connected layer.
-constexpr size_t kFullyConnectedLayersMaxUnits = 24;
+constexpr int kFullyConnectedLayersMaxUnits = 24;
 
 // Maximum number of units for a recurrent layer. This value is used to
 // over-allocate space for recurrent layers state vectors (implemented as
 // std::array). The value should equal the number of units of the largest
 // recurrent layer.
-constexpr size_t kRecurrentLayersMaxUnits = 24;
+constexpr int kRecurrentLayersMaxUnits = 24;
 
 // Fully-connected layer.
 class FullyConnectedLayer {
  public:
-  FullyConnectedLayer(size_t input_size,
-                      size_t output_size,
+  FullyConnectedLayer(int input_size,
+                      int output_size,
                       rtc::ArrayView<const int8_t> bias,
                       rtc::ArrayView<const int8_t> weights,
                       rtc::FunctionView<float(float)> activation_function,
@@ -49,16 +49,16 @@ class FullyConnectedLayer {
   FullyConnectedLayer(const FullyConnectedLayer&) = delete;
   FullyConnectedLayer& operator=(const FullyConnectedLayer&) = delete;
   ~FullyConnectedLayer();
-  size_t input_size() const { return input_size_; }
-  size_t output_size() const { return output_size_; }
+  int input_size() const { return input_size_; }
+  int output_size() const { return output_size_; }
   Optimization optimization() const { return optimization_; }
   rtc::ArrayView<const float> GetOutput() const;
   // Computes the fully-connected layer output.
   void ComputeOutput(rtc::ArrayView<const float> input);
 
  private:
-  const size_t input_size_;
-  const size_t output_size_;
+  const int input_size_;
+  const int output_size_;
   const std::vector<float> bias_;
   const std::vector<float> weights_;
   rtc::FunctionView<float(float)> activation_function_;
@@ -72,8 +72,8 @@ class FullyConnectedLayer {
 // activation functions for the update/reset and output gates respectively.
 class GatedRecurrentLayer {
  public:
-  GatedRecurrentLayer(size_t input_size,
-                      size_t output_size,
+  GatedRecurrentLayer(int input_size,
+                      int output_size,
                       rtc::ArrayView<const int8_t> bias,
                       rtc::ArrayView<const int8_t> weights,
                       rtc::ArrayView<const int8_t> recurrent_weights,
@@ -81,8 +81,8 @@ class GatedRecurrentLayer {
   GatedRecurrentLayer(const GatedRecurrentLayer&) = delete;
   GatedRecurrentLayer& operator=(const GatedRecurrentLayer&) = delete;
   ~GatedRecurrentLayer();
-  size_t input_size() const { return input_size_; }
-  size_t output_size() const { return output_size_; }
+  int input_size() const { return input_size_; }
+  int output_size() const { return output_size_; }
   Optimization optimization() const { return optimization_; }
   rtc::ArrayView<const float> GetOutput() const;
   void Reset();
@@ -90,8 +90,8 @@ class GatedRecurrentLayer {
   void ComputeOutput(rtc::ArrayView<const float> input);
 
  private:
-  const size_t input_size_;
-  const size_t output_size_;
+  const int input_size_;
+  const int output_size_;
   const std::vector<float> bias_;
   const std::vector<float> weights_;
   const std::vector<float> recurrent_weights_;
